added old code back as new code wasn't needed

+++ src/world/generator/populator/Ore.php
@@ -37,9 +37,9 @@ class Ore implements Populator{
 		foreach($this->oreTypes as $type){
 			$ore = new ObjectOre($random, $type);
 			for($i = 0; $i < $ore->type->clusterCount; ++$i){
-				$x = $random->nextRange($chunkX << Chunk::COORD_BIT_SIZE, (int)(($chunkX << Chunk::COORD_BIT_SIZE) + Chunk::EDGE_LENGTH - 1));
+				$x = $random->nextRange($chunkX << Chunk::COORD_BIT_SIZE, ($chunkX << Chunk::COORD_BIT_SIZE) + Chunk::EDGE_LENGTH - 1);
 				$y = $random->nextRange($ore->type->minHeight, $ore->type->maxHeight);
-				$z = $random->nextRange($chunkZ << Chunk::COORD_BIT_SIZE, (int)(($chunkZ << Chunk::COORD_BIT_SIZE) + Chunk::EDGE_LENGTH - 1));
+				$z = $random->nextRange($chunkZ << Chunk::COORD_BIT_SIZE, ($chunkZ << Chunk::COORD_BIT_SIZE) + Chunk::EDGE_LENGTH - 1);
 				if($ore->canPlaceObject($world, $x, $y, $z)){
 					$ore->placeObject($world, $x, $y, $z);
 				}
